pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "redis>=6.4.0",
]
